Lingua::EN::Number::IsOrdinal first release
New module to check if a number is ordinal or cardinal. Has one
function, is_ordinal which returns 1 for ordinal numbers (written or
digits) and undef for cardinal numbers.

Uses Lingua::EN::FindNumber and the regex from perldoc -q 'is a number'
and a simple regex to check for ordinal suffixes.